third_party/blink/web_tests/fast/events/focus-click-on-non-mouse-focusable-element-expected.txt

Click on a focused anchor element should not change focus, and click on a non-focused anchor element should remove focus of another element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is a
PASS document.activeElement is document.body
PASS successfullyParsed is true

TEST COMPLETE

link 
